requirements.txt
# Google APIs
gspread
google-auth
google-auth-oauthlib
google-api-python-client
google-genai

# Data handling
pandas

# RAG / embeddings
sentence-transformers
faiss-cpu

# Utilities
python-dotenv
